Moved action in right package. Minor implementation of action parcelable
